add lazy construct(begin, end) interface to some (mm)phf
... that are required to have this interface upstream. All functions
will be overhauled to support this lazy construction & construction on
presorted range in due time.

sometimes it is desirable to construct an exotic hash function before
the data is available as required, e.g., before it is sorted.

Having a construct() interface will also allow reconstruction during
a function's lifetime.